Bugs removed:  1)GW Recharge value corrected resolved. 2)Cadastral Vulnerability maps rendering issue resolved

+++ kharif_model.py
@@ -282,7 +282,7 @@ class KharifModel:
 			self.input_layers['slope_layer'] = self.iface.addRasterLayer(os.path.join(path, 'Slope.tif'), 'Slope')
 			#~ self.input_layers['drainage_layer'] = self.iface.addRasterLayer(os.path.join(path, 'Drainage.shp'), 'Drainage', 'ogr')
 			
-			self.rain = [int(row["Rainfall"]) for row in csv.DictReader(open(os.path.join(path, RAINFALL_CSV_FILENAME)))]
+			self.rain = [float(row["Rainfall"]) for row in csv.DictReader(open(os.path.join(path, RAINFALL_CSV_FILENAME)))]
 			et0_file_data = [float(row["ET0"]) for row in csv.DictReader(open(os.path.join(path, ET0_CSV_FILENAME)))]
 			self.et0 = set_et0_from_et0_file_data(et0_file_data)
 			self.sowing_threshold = DEFAULT_SOWING_THRESHOLD
@@ -309,7 +309,7 @@ class KharifModel:
 			self.input_layers['slope_layer'] = self.iface.addRasterLayer(self.dlg.slope_layer_filename.text(), 'Slope')
 			if self.dlg.drainage_layer_filename.text() != '':
 				self.drainage_layer = self.iface.addVectorLayer(self.dlg.drainage_layer_filename.text(), 'Drainage', 'ogr')
-			self.rain = [int(row["Rainfall"]) for row in csv.DictReader(open(str(self.dlg.rainfall_csv_filename.text())))]
+			self.rain = [float(row["Rainfall"]) for row in csv.DictReader(open(str(self.dlg.rainfall_csv_filename.text())))]
 			et0_file_data = [float(row["ET0"]) for row in csv.DictReader(open(os.path.join(path, ET0_CSV_FILENAME)))]
 			self.et0 = set_et0_from_et0_file_data(et0_file_data)
 			self.sowing_threshold = self.dlg.sowing_threshold.value()
